requirements.txt
flit~=1.3
click~=7.0
lark-parser~=0.7
pytest~=5.0